+++ App_Data/jobs/continuous/settleGMOAuthorization/run.ts
@@ -1,12 +1,16 @@
 /**
- * ムビチケ着券オーソリキャンセル
+ * GMO実売上
  *
  * @ignore
  */
+
 import * as sskts from '@motionpicture/sskts-domain';
+import * as createDebug from 'debug';
 import * as mongoose from 'mongoose';
 
-import mongooseConnectionOptions from '../../mongooseConnectionOptions';
+import mongooseConnectionOptions from '../../../../mongooseConnectionOptions';
+
+const debug = createDebug('sskts-jobs:bin:watchSettleGMOAuthorizationQueue');
 
 (<any>mongoose).Promise = global.Promise;
 mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
@@ -15,7 +19,7 @@ let count = 0;
 
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
 const INTERVAL_MILLISECONDS = 1000;
-const queueAdapter = sskts.adapter.queue(mongoose.connection);
+const taskAdapter = sskts.adapter.task(mongoose.connection);
 
 setInterval(
     async () => {
@@ -26,7 +30,10 @@ setInterval(
         count += 1;
 
         try {
-            await sskts.service.queue.executeCancelMvtkAuthorization()(queueAdapter);
+            debug('count:', count);
+            await sskts.service.task.executeByName(
+                sskts.factory.taskName.SettleGMOAuthorization
+            )(taskAdapter, mongoose.connection);
         } catch (error) {
             console.error(error.message);
         }