+++ App_Data/jobs/continuous/settleSeatReservationAuthorization/run.ts
@@ -1,12 +1,13 @@
 /**
- * ムビチケ着券オーソリキャンセル
+ * COA仮予約資産移動
  *
  * @ignore
  */
+
 import * as sskts from '@motionpicture/sskts-domain';
 import * as mongoose from 'mongoose';
 
-import mongooseConnectionOptions from '../../mongooseConnectionOptions';
+import mongooseConnectionOptions from '../../../../mongooseConnectionOptions';
 
 (<any>mongoose).Promise = global.Promise;
 mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
@@ -14,8 +15,8 @@ mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
 let count = 0;
 
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
-const INTERVAL_MILLISECONDS = 1000;
-const queueAdapter = sskts.adapter.queue(mongoose.connection);
+const INTERVAL_MILLISECONDS = 500;
+const taskAdapter = sskts.adapter.task(mongoose.connection);
 
 setInterval(
     async () => {
@@ -26,7 +27,9 @@ setInterval(
         count += 1;
 
         try {
-            await sskts.service.queue.executeCancelMvtkAuthorization()(queueAdapter);
+            await sskts.service.task.executeByName(
+                sskts.factory.taskName.SettleSeatReservationAuthorization
+            )(taskAdapter, mongoose.connection);
         } catch (error) {
             console.error(error.message);
         }