+++ bin/watchDisableTransactionInquiryQueue/index.ts
@@ -9,12 +9,12 @@ import * as mongoose from 'mongoose';
 import mongooseConnectionOptions from '../../mongooseConnectionOptions';
 
 (<any>mongoose).Promise = global.Promise;
-mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions);
+mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
 
 let count = 0;
 
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
-const INTERVAL_MILLISECONDS = 250;
+const INTERVAL_MILLISECONDS = 1000;
 const queueAdapter = sskts.adapter.queue(mongoose.connection);
 const transactionAdapter = sskts.adapter.transaction(mongoose.connection);