+++ bin/watchRunningExportTransactionQueue/index.ts
@@ -9,15 +9,15 @@ import * as mongoose from 'mongoose';
 
 import mongooseConnectionOptions from '../../mongooseConnectionOptions';
 
-const debug = createDebug('sskts-api:*');
+const debug = createDebug('sskts-jobs:*');
 
 (<any>mongoose).Promise = global.Promise;
-mongoose.connect(process.env.MONGOLAB_URI, mongooseConnectionOptions);
+mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions);
 
 let countRetry = 0;
 
 const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
-const INTERVAL_MILLISECONDS = 250;
+const INTERVAL_MILLISECONDS = 500;
 const transactionAdapter = sskts.adapter.transaction(mongoose.connection);
 const RETRY_INTERVAL_MINUTES = 10;