+++ CHANGELOG.md
@@ -16,6 +16,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v3.2.0 - 2017-07-04
+### Added
+- 各タスクの状態監視ジョブを新規作成。
+- タスクの中止あるいはリトライジョブを新規作成。
+
+### Changed
+- 取引からタスクをエクスポートするように変更。具体的には、bin/watchClosedTransactionとbin/watchExpiredTransactionの調整。
+- MongoDBのslow queriesレポートに対応してindexを追加。
+- タスクや取引を監視させるインターバル時間(秒)を調整。1インスタンスあたりの頻度をやや下げる対応。
+
+### Fixed
+- タスク実行時のソート条件が間違っていたので修正。
+
 ## v3.1.0 - 2017-06-28
 ### Changed
 - パフォーマンス在庫状況表現を空席率(%)に変更。